The Datasets
The final data was the Empathetic Dialogues Train data filtered using the code elsewhere in this GitHub.

The Test file is a sample of the dataset used. Please search for this github: https://github.com/facebookresearch/EmpatheticDialogues?tab=readme-ov-file
If you want the actual data used in this research project.